location manager now fetches stories given context

+++ Assets/Scripts/Mascot.cs
@@ -52,7 +52,7 @@ public class Mascot : MonoBehaviour
         return mascotName;
     }
 
-    public float GetHeartLevel()
+    public int GetHeartLevel()
     {
         return heartLevel;
     }
